reshuffling code to solve #299
 Also set verbose=False in time_unit_to_datum_exp_dir( ) to avoid incessant notifications.
