+++ pyleoclim/utils/tsutils.py
@@ -19,7 +19,6 @@ __all__ = [
     'remove_outliers'
 ]
 
-import operator
 import numpy as np
 import pandas as pd
 import warnings
@@ -44,125 +43,6 @@ from .tsbase import (
     clean_ts
 )
 
-SECONDS_PER_YEAR = 365.25 * 60  * 60 * 24
-
-MATCH_A  = frozenset(['y', 'yr', 'yrs', 'year', 'years'])
-MATCH_KA = frozenset(['ka', 'ky', 'kyr', 'kyrs', 'kiloyear', 'kiloyr', 'kiloyrs']) 
-MATCH_MA = frozenset(['ma', 'my','myr','myrs'])
-MATCH_GA = frozenset(['ga', 'gy', 'gyr', 'gyrs'])
-
-def time_unit_to_datum_exp_dir(time_unit, time_name=None, verbose=True):
-         
-    tu = time_unit.lower().split()
-    
-    # deal with statements explicit about exponents, and take a first guess at datum/direction        
-    if tu[0] in MATCH_A:
-        exponent = 0  
-        datum = 0
-        direction = 'prograde'
-    elif tu[0] in MATCH_KA:
-        datum = 1950
-        exponent = 3
-        direction = 'retrograde'
-    elif tu[0] in MATCH_MA:
-        datum = 1950
-        exponent = 6
-        direction = 'retrograde'
-    elif tu[0] in MATCH_GA:
-        datum = 1950
-        exponent = 9
-        direction = 'retrograde'
-    elif tu[0].replace('.','') in ['ad', 'ce']:
-        exponent = 0  
-        datum = 0
-        direction = 'prograde'    
-    else:
-        warnings.warn(f'Time unit "{time_unit}" unknown; triggering defaults', stacklevel=4)    
-        exponent = 0  
-        datum = 0
-        direction = 'prograde'
-    
-    # if provided, deal with statements about datum/direction, like kyr BP, years CE, etc
-    if len(tu) > 1: 
-        datum_str = tu[1].replace('.','') # make lowercase + strip stops, so "B.P." --> "bp"
-        if datum_str == 'b2k':
-            datum = 2000
-            direction = 'retrograde'
-        elif datum_str in ['bp', 'bnf', 'b1950']:
-            datum = 1950
-            direction = 'retrograde'
-        elif datum_str in ['ad', 'ce']:
-            datum = 0
-            direction = 'prograde'
-    
-    if time_name is not None:
-        if time_name.lower() == 'age':
-            direction = 'retrograde'
-        
-    if verbose:
-        print(f'Provided time medata translated to {direction} flow, 10^{exponent} year units, and year {datum} datum')    
-  
-    return (datum, exponent, direction)
-
-def convert_datetime_index_to_time(datetime_index, time_unit, time_name):
-    datum, exponent, direction = time_unit_to_datum_exp_dir(time_unit, time_name)
-    #import operator
-    if direction == 'prograde':
-        multiplier = 1
-    elif direction == 'retrograde':
-        multiplier = -1
-    else:
-        raise ValueError(f'Expected one of {"prograde", "retrograde"}, got {direction}')
-        
-    if not isinstance(datetime_index, pd.DatetimeIndex): 
-        raise ValueError('The provided index is not a proper DatetimeIndex object')
-    else:
-        year_diff = (datetime_index.year - int(datum))
-        numpy_datetime_index = datetime_index.to_numpy()
-        years_floor = numpy_datetime_index.astype('datetime64[Y]').astype('datetime64[s]')
-        seconds_diff = (numpy_datetime_index - years_floor).astype('int')
-        diff = year_diff + seconds_diff / SECONDS_PER_YEAR
-        time = multiplier * diff / 10**exponent
-
-    return time
-
-def time_to_datetime(time, datum=0, exponent=0, direction='prograde', unit='s'):
-    '''
-    Converts a vector of time values to a pandas datetime object
-
-    Parameters
-    ----------
-    time : array-like
-        the time axis to be converted
-    datum : int, optional
-        origin point for the time scale. The default is 0.
-    exponent : int, optional
-        Base-10 exponent for year multiplier. Dates in kyr should use 3, dates in Myr should use 6, etc.
-        The default is 0. 
-    direction : str, optional
-        Direction of time flow, 'prograde' [default] or 'retrograde'.
-    unit : str, optional
-        Units of the datetime. Default is 's', corresponding to seconds. 
-        Only change if you have an excellent reason to use finer resolution!
-
-    Returns
-    -------
-    index, a datetime64[unit] object
-
-    '''
-    if direction == 'prograde':
-        op = operator.add
-    elif direction == 'retrograde':
-        op = operator.sub
-    else:
-        raise ValueError(f'Expected one of {"prograde", "retrograde"}, got {direction}')
-    
-    timedelta = np.array(time) * 10**exponent
-    years = timedelta.astype('int')
-    seconds = ((timedelta - timedelta.astype('int')) * SECONDS_PER_YEAR).astype('timedelta64[s]') # incorporate unit here
-    index = op(op(int(datum), years).astype(str).astype('datetime64[s]'), seconds)  # incorporate unit here?
-    
-    return index
 
 def simple_stats(y, axis=None):
     """ Computes simple statistics
